Rename to agent utils and remove check_structured_output helper
